feat: added backing field for TokenTransactions for improving complex queries feat: added total token spent to Workflow GET request

+++ api/RAGNet.Domain/TokenWallets/TokenTransactions/TransactionDailyAggregation.cs
@@ -3,7 +3,7 @@ namespace RAGNET.Domain.TokenWallets.TokenTransactions
     public class TransactionDailyAggregation
     {
         public DateTime Date { get; set; }
-        public long FreeTokensConsumed { get; set; }
-        public long PaidTokensConsumed { get; set; }
+        public double FreeTokensConsumed { get; set; }
+        public double PaidTokensConsumed { get; set; }
     }
 }
